+++ client/react-utils/filter-input/action-creators.js
@@ -1,0 +1,4 @@
+import actions from './actions';
+import actionCreator from './../action-creator';
+
+export const updateValue = (id, value) => actionCreator(actions.UPDATE_VALUE, { id, value });