+++ client/react-utils/filter-input/namespace.js
@@ -1,0 +1,3 @@
+import namespace from './../namespace';
+
+export default (key) => namespace(`FILTER_INPUT.${key}`);