+++ client/react-utils/reducers.js
@@ -1,0 +1,6 @@
+import filterInputReducers from './filter-input/reducers';
+import concatenateReducers from './concatenate-reducers';
+
+export default concatenateReducers([
+    filterInputReducers
+]);